requirements.txt
jsonschema
loguru
pyyaml
flask